requirements.txt
numpy
networkx==3.4.2
gymnasium==1.2.3
pyyaml==6.0.3
pandas==2.3.3
matplotlib==3.10.8
scipy==1.15.3
torch==2.9.1
tqdm==4.60.0
python-dotenv==1.2.1
pydantic_settings==2.12.0
pydantic==2.12.5
streamlit==1.52.2
pytest==8.4.2